requirements.txt
flask==2.3.3
flask-cors==4.0.0
crewai==0.28.8
crewai-tools==0.1.6
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
groq==0.9.0
python-dotenv==1.0.0
pydantic==2.5.3